feat: Add landing page and update README.md
